Scrolling Background
So player moves like normal but we call updateCamera() now so that we have a scrolling camera type of system now. updateCamera() find's player's x position and figures out how much we need to move world so player looks centered. Then we have the world.setLayoutX(-cameraX) to move the world. Also added another platform on far edge to test if it scrolls right.